# farin-main/docs/swagger.yaml
definitions:
...
  farin_app_api_response.Response-farin_domain_dto_DriverListResponse:
    properties:
      data:
        $ref: '#/definitions/farin_domain_dto.DriverListResponse'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_app_api_response.Response-farin_domain_dto_DriverResponse:
    properties:
      data:
        $ref: '#/definitions/farin_domain_dto.DriverResponse'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_app_api_response.Response-swagger_EmptyObject:
    properties:
      data:
        $ref: '#/definitions/swagger.EmptyObject'
      message:
        type: string
      status:
        type: integer
    type: object
...
  farin_domain_dto.DriverListResponse:
    properties:
      drivers:
        items:
          $ref: '#/definitions/farin_domain_dto.DriverResponse'
        type: array
      total:
        type: integer
    type: object
...
  farin_domain_dto.DriverRequest:
    properties:
      birthCertificateImage:
        items:
          type: integer
        type: array
      contractorId:
        type: string
      criminalRecordImage:
        items:
          type: integer
        type: array
      description:
        type: string
      driverPhoto:
        items:
          type: integer
        type: array
      driverType:
        enum:
        - original
        - reserve
        type: string
      employmentEndDate:
        type: integer
      employmentStartDate:
        type: integer
      employmentStatus:
        type: string
      healthCertificateImage:
        items:
          type: integer
        type: array
      idcardImage:
        items:
          type: integer
        type: array
      militaryServiceCardImage:
        items:
          type: integer
        type: array
      shiftType:
        enum:
        - morning
        - afternoon
        - both
        type: string
      userId:
        type: string
    required:
    - contractorId
    - driverType
    - employmentStartDate
    - employmentStatus
    - shiftType
    - userId
    type: object
  farin_domain_dto.DriverResponse:
    properties:
      contractorId:
        type: string
      createdAt:
        type: integer
      description:
        type: string
      driverType:
        type: string
      employmentEndDate:
        type: integer
      employmentStartDate:
        type: integer
      employmentStatus:
        type: string
      id:
        type: string
      shiftType:
        type: string
      updatedAt:
        type: integer
      userId:
        type: string
    type: object
...
  swagger.EmptyObject:
    type: object
info:
  contact: {}
paths:
...
  /drivers:
    get:
      consumes:
      - application/json
      description: Retrieve a list of drivers with optional filters and pagination
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - default: created_at
        description: Field to sort by
        in: query
        name: sortField
        type: string
      - default: asc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: List drivers
      tags:
      - drivers
    post:
      consumes:
      - application/json
      description: Create a new driver by providing driver details
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Driver Data
        in: body
        name: driver
        required: true
        schema:
          $ref: '#/definitions/farin_domain_dto.DriverRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Create a new driver
      tags:
      - drivers
  /drivers/{id}:
    delete:
      description: Delete a driver by ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Delete a driver
      tags:
      - drivers
    get:
      description: Retrieve driver details by ID
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Get driver details
      tags:
      - drivers
    put:
      consumes:
      - application/json
      description: Update an existing driver's details
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated Driver Data
        in: body
        name: driver
        required: true
        schema:
          $ref: '#/definitions/farin_domain_dto.DriverRequest'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-farin_domain_dto_DriverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/farin_app_api_response.Response-swagger_EmptyObject'
      security:
      - ApiKeyAuth: []
      summary: Update driver details
      tags:
      - drivers
...
swagger: "2.0"